Move url templates to separate class.
