+++ Hqub.MusicBrainz/Hqub.MusicBrainz.API/MusicBrainzClient.cs
@@ -23,10 +23,6 @@
         /// </summary>
         public const string ServiceBaseAddress = "https://musicbrainz.org/ws/2/";
 
-        private const string LookupTemplate = "{0}/{1}/?inc={2}&fmt=json";
-        private const string BrowseTemplate = "{0}?{1}={2}&limit={3}&offset={4}&inc={5}&fmt=json";
-        private const string SearchTemplate = "{0}?query={1}&limit={2}&offset={3}&fmt=json";
-
         private const string UserAgent = "Hqub.MusicBrainz/3.0-beta";
 
         #region Public services
@@ -98,11 +94,13 @@
         /// <param name="proxy">The <see cref="IWebProxy"/> used to connect to the webservice.</param>
         public MusicBrainzClient(string baseAddress, IWebProxy proxy)
         {
-            Artists = new ArtistService(this);
-            Recordings = new RecordingService(this);
-            Releases = new ReleaseService(this);
-            ReleaseGroups = new ReleaseGroupService(this);
-            Work = new WorkService(this);
+            var urlBuilder = new UrlBuilder(true);
+
+            Artists = new ArtistService(this, urlBuilder);
+            Recordings = new RecordingService(this, urlBuilder);
+            Releases = new ReleaseService(this, urlBuilder);
+            ReleaseGroups = new ReleaseGroupService(this, urlBuilder);
+            Work = new WorkService(this, urlBuilder);
 
             client = CreateHttpClient(new Uri(baseAddress), true, proxy);
         }
@@ -160,102 +158,6 @@
 
             return new WebServiceException(error.Message, status, url);
         }
-
-        #region Generate urls
-
-        /// <summary>
-        /// Creates a webservice lookup template.
-        /// </summary>
-        internal string CreateLookupUrl(string entity, string mbid, params string[] inc)
-        {
-            return CreateLookupUrl(entity, mbid, string.Join("+", inc));
-        }
-
-        /// <summary>
-        /// Creates a webservice lookup template.
-        /// </summary>
-        internal string CreateLookupUrl(string entity, string mbid, string inc)
-        {
-            return string.Format(LookupTemplate, entity, mbid, inc);
-        }
-
-        /// <summary>
-        /// Creates a webservice browse template.
-        /// </summary>
-        internal string CreateBrowseTemplate(string entity, string relatedEntity, string mbid, int limit, int offset, params string[] inc)
-        {
-            return CreateBrowseTemplate(entity, relatedEntity, mbid, limit, offset, string.Join("+", inc));
-        }
-
-        /// <summary>
-        /// Creates a webservice browse template.
-        /// </summary>
-        internal string CreateBrowseTemplate(string entity, string relatedEntity, string mbid, string type, string status,
-            int limit, int offset, params string[] inc)
-        {
-            var url = CreateBrowseTemplate(entity, relatedEntity, mbid, limit, offset, string.Join("+", inc));
-
-            if (!ValidateBrowseParam(Resources.Constants.BrowseStatus, status))
-            {
-                throw new ArgumentException(string.Format(Resources.Messages.InvalidQueryValue, status, "status"));
-            }
-
-            if (!ValidateBrowseParam(Resources.Constants.BrowseType, type))
-            {
-                throw new ArgumentException(string.Format(Resources.Messages.InvalidQueryValue, type, "type"));
-            }
-
-            if (!string.IsNullOrEmpty(type))
-            {
-                url += "&type=" + type;
-            }
-
-            if (!string.IsNullOrEmpty(status))
-            {
-                url += "&status=" + status;
-            }
-
-            return url;
-        }
-
-        /// <summary>
-        /// Creates a webservice browse template.
-        /// </summary>
-        internal string CreateBrowseTemplate(string entity, string relatedEntity, string mbid, int limit, int offset, string inc)
-        {
-            return string.Format(BrowseTemplate, entity, relatedEntity, mbid, limit, offset, inc);
-        }
-
-        /// <summary>
-        /// Creates a webservice search template.
-        /// </summary>
-        internal string CreateSearchTemplate(string entity, string query, int limit, int offset)
-        {
-            query = Uri.EscapeUriString(query);
-
-            return string.Format(SearchTemplate, entity, query, limit, offset);
-        }
-
-        private bool ValidateBrowseParam(string availableParams, string value)
-        {
-            if (string.IsNullOrEmpty(value))
-            {
-                return true; // Irgnore, if no value specified.
-            }
-
-            if (value.IndexOf('|') > 0)
-            {
-                return value.Split('|').All(s => availableParams.IndexOf("+" + s + "+") >= 0);
-            }
-
-            return availableParams.IndexOf("+" + value + "+") >= 0;
-        }
-
-        #endregion
-
-        // TODO: should HttpClient be re-used?
-
-        // https://medium.com/@nuno.caneco/c-httpclient-should-not-be-disposed-or-should-it-45d2a8f568bc
 
         private HttpClient CreateHttpClient(Uri baseAddress, bool automaticDecompression, IWebProxy proxy)
         {
